+++ test/environments/autodiff/src/PureRNGsAutodiffConformance.jl
@@ -1,0 +1,1 @@
+module PureRNGsAutodiffConformance end